Hide engine implementation from user and add support for engine disabling: - base class Engine returned to user (corresponds to disabled engine) - derived class EngineImpl (corresponds to enabled engine)
